Move provider in the same file
