--- Mastodon/Scene/Profile/Following/FollowingListViewController+DataSourceProvider.swift
@@ -1,34 +1,0 @@
-//
-//  FollowingListViewController+DataSourceProvider.swift
-//  Mastodon
-//
-//  Created by MainasuK on 2022-1-20.
-//
-
-import UIKit
-
-extension FollowingListViewController: DataSourceProvider {
-    func item(from source: DataSourceItem.Source) async -> DataSourceItem? {
-        var _indexPath = source.indexPath
-        if _indexPath == nil, let cell = source.tableViewCell {
-            _indexPath = await self.indexPath(for: cell)
-        }
-        guard let indexPath = _indexPath else { return nil }
-        
-        guard let item = viewModel.diffableDataSource?.itemIdentifier(for: indexPath) else {
-            return nil
-        }
-        
-        switch item {
-            case .account(let account, let relationship):
-                return .account(account: account, relationship: relationship)
-            default:
-                return nil
-        }
-    }
-    
-    @MainActor
-    private func indexPath(for cell: UITableViewCell) async -> IndexPath? {
-        return tableView.indexPath(for: cell)
-    }
-}
