+++ Mastodon/Scene/Profile/Following/FollowingListViewController.swift
@@ -126,3 +126,31 @@ extension FollowingListViewController: UITableViewDelegate, AutoGenerateTableVie
 
 // MARK: - UserTableViewCellDelegate
 extension FollowingListViewController: UserTableViewCellDelegate {}
+
+
+// MARK: - DataSourceProvider
+extension FollowingListViewController: DataSourceProvider {
+    func item(from source: DataSourceItem.Source) async -> DataSourceItem? {
+        var _indexPath = source.indexPath
+        if _indexPath == nil, let cell = source.tableViewCell {
+            _indexPath = await self.indexPath(for: cell)
+        }
+        guard let indexPath = _indexPath else { return nil }
+
+        guard let item = viewModel.diffableDataSource?.itemIdentifier(for: indexPath) else {
+            return nil
+        }
+
+        switch item {
+            case .account(let account, let relationship):
+                return .account(account: account, relationship: relationship)
+            default:
+                return nil
+        }
+    }
+
+    @MainActor
+    private func indexPath(for cell: UITableViewCell) async -> IndexPath? {
+        return tableView.indexPath(for: cell)
+    }
+}
